feat(qbittorrent): dont chown/setuid in entrypoint
Instead of a complicated entrypoint, we introduce a coreutils container;
this will run as root before the application allowing us to
chown/setuid/etc. any mounted volumes to get them ready for the
application.

This way the application can simply run directly as whatever user is
passed in instead of starting as root and then setuid based on an
environment variable.