build(pbr): specify tag to not require rebuild
Without a tag, buildLayeredImage will use the hash of the built
artifact, which can not be known until after building. This then
introduces pbr as a build depencency to the manifest; the end result is
it slows down top-level builds.

Since the tag isn't actually important, we can hard-code it to speed
things up.